Fix bug triggering exit warning in __fasd_run
If the 'exit' command was part of the command line, then fish would
give a warning and not exit due to how the fasd command was running
in the background.

Adding a disown after it will stop it from throwing this error.

Fix small issue with printf formatting in __fasd_expand_vars